fix: uses of `ucal_getHostTimeZone` on systems that don't have it
The compile-time Windows version check is not a reliable way of checking
whether the system has the `ucal_getHostTimeZone` function in its
icu.dll. That function is only available since ICU 65, which only ships
on Windows 11 and newer. Supporting that is not worth the hassle, so we
keep the "manual" version (for now).